Move the buffered logger to internal because the rosetta stuff needed it (was included in the non-unit-test-only app/test_helpers.go file) but couldn't get it because it wasn't part of accessible stuff.
