+++ internal/logger.go
@@ -1,4 +1,4 @@
-package mocks
+package internal
 
 import (
 	"bytes"
